requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
Deprecated==1.3.1
diff-match-patch==20241021
dj-database-url==3.1.2
Django==6.0.3
django-import-export==4.4.0
finance-datareader==0.9.110
fonttools==4.62.1
frozendict==2.4.7
idna==3.11
kiwisolver==1.5.0
llvmlite==0.46.0
lxml==6.0.2
matplotlib==3.10.8
multipledispatch==1.0.0
multitasking==0.0.12
narwhals==2.18.1
numba==0.64.0
numpy==1.26.4
packaging==26.0
pandas==2.3.3
peewee==4.0.2
pillow==12.2.0
platformdirs==4.9.4
plotly==6.6.0
protobuf==7.34.0
psycopg2-binary==2.9.11
pycparser==3.0
pykrx==1.2.4
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
requests-file==3.0.1
schedule==1.2.2
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
tablib==3.9.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
websockets==16.0
whitenoise==6.12.0
wrapt==2.1.2
yfinance==1.2.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.0